requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
joblib==1.4.0
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
PyMySQL==1.1.0
scikit_learn==1.3.2
bcrypt==4.3.0
PyJWT==2.9.0
